requirements.txt
astroid==2.11.2
attrs==21.4.0
boto3==1.21.32
botocore==1.24.32
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
dill==0.3.4
Flask==2.1.1
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
moto==3.1.3
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.13.4
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
responses==0.20.0
s3transfer==0.5.2
six==1.16.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
xmltodict==0.12.0
zipp==3.7.0
